backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
qdrant-client==1.9.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
cohere==5.5.4
aiohttp==3.9.5
pytest==7.4.3